+++ packages/react-querybuilder/src/internal/generateID.test.ts
@@ -1,8 +1,0 @@
-import { generateID } from './generateID';
-
-it('should generate unique IDs', () => {
-  const id1 = generateID();
-  const id2 = generateID();
-
-  expect(id1).not.toBe(id2);
-});