+++ packages/react-querybuilder/src/internal/index.ts
@@ -1,5 +1,4 @@
 export * from './filterFieldsByComparator';
-export * from './generateID';
 export * from './getValidationClassNames';
 export * from './getValueSourcesUtil';
 export * from './regenerateIDs';